+++ src/applications/health-care-questionnaire/questionnaire/containers/ConfirmationPage/index.jsx
@@ -8,7 +8,7 @@ import {
 import ConfirmationPageFooter from '../../components/confirmation-page-footer/ConfirmationPageFooter';
 import AppointmentDisplay from '../../components/appointment-display/AppointmentDisplay';
 
-import PrintButton from '../../../list/questionnaire-list/components/Shared/Print/PrintButton';
+import PrintButton from '../../../shared/components/print/PrintButton';
 
 const ConfirmationPage = props => {
   const { appointment } = props;
