+++ src/applications/health-care-questionnaire/shared/components/print/ViewAndPrint.jsx
